ROE-1646 Introduce abstract class that encapsulates common logic
